Extend Calculator with square and square root methods
